refactor(chezwiz): remove LLMProvider abstraction
Simplify provider architecture by removing the intermediate LLMProvider trait.
The OpenAICompatibleProvider now works standalone without the abstraction layer,
making the codebase more direct and easier to maintain.
